+++ robot/strategy/gameTargetActionList.h
@@ -36,7 +36,7 @@ void addTargetAction(GameTargetActionList* targetActionList,
                      GameTargetAction* targetAction,
                      Location* startLocation,
                      Location* endLocation,
-                     int timeToAchieve,
+                     float timeToAchieve,
                      PathData* pathData,
                      GameTargetActionItemList* actionListItem
 );
